Added support for partially custom client messages
This allows the client to send stuff like mouse clicks, keyboard presses, etc. in abstracted form to the server.  The client maps the `ClientEvent` which is either mouse click, keyboard press, or some other custom event (we can extend this in the future easily) into a `ClientMessage<T>` which the server then handles given a &mut Universe and the option to broadcast a message to all connected clients or back to the client that initiated the custom event.